requirements.txt
fastapi
uvicorn
httpx
pydantic[email]
email-validator
google-auth
google-api-python-client